Asianelephants Dominance (unweighted)

Species: Elephas maximus

Taxonomic class: Mammalia

Population type: free-ranging

Geographical location: Uda Walawe National Park, Sri Lanka

Data collection technique: focal sampling

Interaction type: dominance

Definition of interaction: Indicators of dominance as well as subordination was included. If a series of interactions occurred during a particular event, the winners/losers were determined only on conclusion of the event, when individuals or groups moved apart.

Edge weight type: unweighted

Total duration of data collection: 206days

Time resolution of data collection (within a day): 1sec

Time span of data collection (within a day): 5.5 hours

Note:

Usage

animal_15

Format

list of igraph objects

Source

https: //bansallab.github.io/asnr/

References

de Silva, Shermin, Volker Schmid, and George Wittemyer. "Fission–fusion processes weaken dominance networks of female Asian elephants in a productive habitat." Behavioral Ecology (2016): arw153.