Rhesusmacaque Association (weighted)

Species: Macaca mulatta

Taxonomic class: Mammalia

Population type: captive

Geographical location: Rijswijk, the Netherlands

Data collection technique: survey scan

Interaction type: physical contact

Definition of interaction: Edge defined based on individuals sitting in contact with eah other.

Edge weight type: frequency

Total duration of data collection: 3years

Time resolution of data collection (within a day): 1 hour

Time span of data collection (within a day): few minutes

Note: Two networks represent proximity (contact sits) and grooming patterns - who was sitting in contact with whom and who was grooming whom.

Usage

animal_29

Format

list of igraph objects

Source

https: //bansallab.github.io/asnr/

References

Massen, Jorg JM, and Elisabeth HM Sterck. "Stability and durability of intra-and intersex social bonds of captive rhesus macaques (Macaca mulatta)." International Journal of Primatology 34.4 (2013): 770-791.