Jemaah Islamiyah Koschade

Jemaah Islamiyah cell that was responsible for the Bali bombings in 2002 – i.e. should tally with the other Bali bombing dataset from JJATT. The recording of the interaction of the cell began following the meeting in the Hotel Harem in Denpasar on October 6, when the group was considered to go ‘operationally covert’, and concluded when the majority of the group had left Bali before the implementation of the operation on October 11, 2002. 1-mode matrix 17 x 17 person by person, undirected and valued. Relationships between terrorists concern who exchanged information with whom (communications exchanges). The valued relations represent the strength of the relations between the individuals, with a score of one signifying the weakest relationship such as a single text message or a financial transaction, and five signifying the strongest relationship such as individuals who resided together, or individuals who had numerous weak contacts over the period in question.

Usage

covert_26

Format

igraph object

Source

Dataset comes from the publication Koschade, Stuart (2006) A Social Network Analysis of Jemaah Islamiyah: The Applications to Counter-Terrorism and Intelligence. Studies in Conflict and Terrorism Vol. 29(6):pp. 559-575 This data has been reconstructed by Koschade by using the report provided by the international crisis group which collected depositions of JI suspects, court documents, and others Indonesian press reports. (INTERNATIONAL CRISIS GROUP. 2003 ‘Jemaah Islamiyah in South East Asia: Damaged but Still Dangerous’).