London Gang

Data is on co-offending in a London-based inner-city street gang, 2005-2009, operating from a social housing estate. Data comes from anonymised police arrest and conviction data for ‘all confirmed’ members of the gang. 1-Mode matrix 54 x 54 persons by persons, undirected, valued.

Network tie values:
               = 1 (hang out together)
               = 2 (co-offend together)
               = 3 (co-offend together, serious crime)
               = 4 (co-offend together, serious crime, kin)

Attributes: Age, Birthplace (1 = West Africa, 2= Caribbean, 3= UK, 4= East Africa), Residence, Arrests, Convictions, Prison, Music.

Usage

covert_28

Format

igraph object

Source

Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks).

References

Grund, T. and Densley, J. (2015) Ethnic Homophily and Triad Closure: Mapping Internal Gang Structure Using Exponential Random Graph Models. Journal of Contemporary Criminal Justice, Vol. 31, Issue 3, pp. 354-370 Grund, T. and Densley, J. (2012) Ethnic Heterogeneity in the Activity and Structure of a Black Street Gang. European Journal of Criminology, Vol. 9, Issue 3, pp. 388-406.