Provisional Irish Republican Army

Data is on active Provisional IRA (hereafter PIRA) members between 1970 and 1998. Data collected at the International Center for the Study of Terrorism, Pennsylvania State University. From this data are derived network structures and the nature of dependence within them. The PIRA network comprises the following four types of relationships: (1) involvement in a PIRA activity together, (2) friends before joining PIRA movement, (3) blood relatives, and (4) related through marriage. We treated each relation as a tie and coded whether a tie exists between two members or not. Thus, the networks have, conceptually and technically, binary and symmetric relations between members. Data also includes sociological information of members, such as gender, age, marital status, recruiting age, education (that is, attending university), brigade memberships, non-/violent characteristics, role-related characteristics—senior leader, IED constructor, IED planter, and gunman—and task-related characteristics (that is, foreign operation tasks, and involvement n bank robbery, kidnapping, hijacking, and drugs).

Usage

covert_38

Format

list of igraph objects

Source

Available from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks).

References

Paul Gill , Jeongyoon Lee , Karl R. Rethemeyer , John Horgan & Victor Asal (2014) Lethal Connections: The Determinants of Network Connections in the Provisional Irish Republican Army, 1970–1998, International Interactions: Empirical and Theoretical Research in International Relations, 40:1, 52-78