Baseball Steroid Use

“When the Mitchell Report on steroid use in Major League Baseball (MLB), was published people were surprised at both the number and names of those who were mentioned. The diagram below shows a network map created from data found in the Mitchell Report. Baseball players are shown as green nodes. Those who were found to be providers of steroids and other illegal performance enhancing substances appear as red nodes. The links reveal the flow of chemicals – from providers to players.” http://orgnet.com/steroids.html 2-mode matrix 72 x 12 of users by providers, relations are the supply of illegal performance enhancing substances. 1-mode matrix 72x72 of player by player, relations are having suppliers in common.

Usage

covert_8

Format

list of igraph objects

Source

Freely available http://orgnet.com/steroids.html and reconstructed by Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks)