Big Allied and Dangerous (BAAD)

The Big Allied And Dangerous (BAAD) project focuses on creation and maintenance of a comprehensive database of terrorist organizational characteristics. This project has created two datasets. The first, BAAD Version 1.0 (BAAD1) contains a single snapshot of 395 terrorist organizations active (meaning they perpetrated at least one attack) between 1998-2005. This dataset grew from the information originally hosted by the Memorial Institute for the Prevention of Terrorism’s (MIPT) in their Terrorism Knowledge Base (TKB). BAAD1 improved and extended the data available from MIPT through independent verification and coding efforts. The dataset includes both case-by-variables information on each organization and network data. The case-by-variables dataset is available for download currently. The network data will be available for download during the second quarter of 2010. Big Allied and Dangerous, Version 2.0 (BAAD2) seeks to improve upon BAAD1 in multiple ways by (1) providing time series data in yearly slices, (2) expanding the time period forward through 2007, and (3) increasing the number and depth of variables collected and coded. BAAD2 is made up of two major components. The first is the data on organizational variables. These variables include: group name, aliases, homebase, ideology, size, age, structure, financial support, electoral involvement, leadership loss, territorial control, CBRN pursuit or use, and number of incidents, injuries, and fatalities. The second component is the social network data, which characterizes relations between terrorist organizations as well as between countries and terrorist organizations. Relationships are coded for categories such as: suspected ally, ally, faction, splinter group, rival, enemy, target, and state sponsor. This data can then be used to create dynamic network visualizations to show the networks evolving over the 10 years included in the dataset. Data construction for BAAD2 is currently ongoing. NETWORK 2-mode matrix 394 x 65 organization by territory, undirected ties. Ties are location of attacks. 1-mode matrix 394 x 394 organization by organization (co-location of attacks), undirected ties. Attribute data for each organization. Attribute codebook available at: http://www.albany.edu/pvc/lethality_paper__CodeBook.pdf

Usage

covert_9

Format

list of igraph objects

Source

Freely available from http://www.albany.edu/pvc/data.shtml and from Manchester (https://sites.google.com/site/ucinetsoftware/datasets/covert-networks)

References

Asal, Victor H. and R. Karl Rethemeyer. (2008). “The Nature of the Beast: Terrorist Organizational Characteristics and Organizational Lethality.” Journal of Politics, 70(2): 437-449.