EIES (relations)

These data arose from an early experiment on computer mediated communication. Fifty academics interested in social network research were allowed to contact each other via an Electronic Information Exchange System (EIES). The data collected consisted of all messages sent plus acquaintance relationships at two time periods (collected via a questionnaire).The data include the 32 actors who completed the study. In addition attribute data on primary discipline and number of citations was recorded.

TIME_1 and TIME_2 give the reported acquaintance information at the beginning of the study and eight months later. These are coded as follows: 4 = close personal fiend, 3= friend, 2= person I've met, 1 = person I've heard of but not met, and 0 = person unknown to me (or no reply).

The attribute data give the number of citations of the actors work in the social science citation index at the beginning of the study together with a discipline code: 1 = Sociology, 2 = Anthropology, 3 = Mathematics/Statistics, 4 = other. These data are used by Wasserman and Faust in their network analysis book.

Usage

eies_relations

Format

igraph object

Source

http://moreno.ss.uci.edu/data.html#eies

References

Freeman, S. C. and L. C. Freeman (1979). The networkers network: A study of the impact of a new communications medium on sociometric structure. Social Science Research Reports No 46. Irvine CA, University of California.

Wasserman S. and K. Faust (1994). Social Network Analysis: Methods and Applications.Cambridge University Press, Cambridge.

See also