Newcomb Fraternity

The networksrecord weekly sociometric preference rankings from 17 men attending the University of Michigan in the fall of 1956; data from week 9 are missing. A "1" indicates first preference, and no ties were allowed.

The men were recruited to live in off-campus (fraternity) housing, rented for them as part of the Michigan Group Study Project supervised by Theodore Newcomb from 1953 to 1956. All were incoming transfer students with no prior acquaintance of one another. After the meeting, Lin Freeman cleaned the data set and made it available on his website. It takes the form of a matrix that records coauthorship among 475 authors who were involved in the production of 295 articles. Cell entries report the number of coaurherships displayed by pairs of authors.

Usage

fraternity

Format

list of 15 igraph objects

Source

http://moreno.ss.uci.edu/data#newfrat

References

Newcomb T. (1961). The acquaintance process. New York: Holt, Reinhard & Winston.

Nordlie P. (1958). A longitudinal study of interpersonal attraction in a natural group setting. Unpublished doctoral dissertation, University of Michigan.

White H., Boorman S. and Breiger R. (1977). Social structure from multiple networks, I. Blockmodels of roles and positions. American Journal of Sociology, 81, 730-780.