High-tech Managers (Reports to)

These data were collected from the managers of a high-tec company. The company manufactured high-tech equipment on the west coast of the United States and had just over 100 employees with 21 managers. Each manager was asked "To whom do you go to for advice?" and "Who is your friend?" Data for the item "To whom do you report?" was taken from company documents. In addition attribute information was collected. This consisted of the managers age (in years), length of service or tenure (in years), level in the corporate hierarchy (coded 1,2 and 3; 1=CEO, 2 = Vice President, 3 = manager) and department (coded 1,2,3,4 with the CEO in department 0 ie not in a department).

Usage

ht_reports

Format

igraph object

Source

http://moreno.ss.uci.edu/data#krackht

References

Krackhardt D. (1987). Cognitive social structures. Social Networks, 9, 104-134.

See also