Netscience Coauthorship

coauthorship network of scientists working on network theory and experiment, as compiled by Mark Newman in May 2006. The network was compiled from the bibliographies of two review articles on networks, M. E. J. Newman, SIAM Review 45, 167-256 (2003) and S. Boccaletti et al., Physics Reports 424, 175-308 (2006), with a few additional references added by hand. The version given here contains all components of the network, for a total of 1589 scientists, and not just the largest component of 379 scientists previously published. The network is weighted, with weights assigned directly in terms of the number of collaborations between authors and inversely in terms of the number of other authors involved. This weighting is described in M. E. J. Newman, Phys. Rev. E 64, 016132 (2001).

Usage

netsci

Format

igraph object

Source

http://moreno.ss.uci.edu/data#netsci

References

M. E. J. Newman, Finding community structure in networks using the eigenvectors of matrices, Preprint physics/0605087 (2006).