networkdata: A repository of network datasets

The package contains a large collection of network datasets with different context.

Details

The package includes data from the following sources:

freeman datasets from http://moreno.ss.uci.edu/data

movie networks from https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/T4HBA3

covert networks from https://sites.google.com/site/ucinetsoftware/datasets/covert-networks

animal networks from https://bansallab.github.io/asnr/

shakespeare networks build with data from https://github.com/mallaham/Shakespeare-Plays

some networks from http://konect.uni-koblenz.de/

- Tennis networks compiled from https://github.com/JeffSackmann (please give credit to him if you use this data)

and a range of miscellaneous network datasets.

How to use the datasets

After loading the networkdata package, the datasets can be loaded using the data command.

Use

  data(package = "networkdata")

to get the complete list of datasets included in this package.

See also

Author

Maintainer: David Schoch david@schochastics.net (ORCID)