Protein-Protein Interactions (probably yeast)

One research area in biology in which centralities have been applied is protein-protein interaction. Interactions between proteins are common. They play an important part in every process involving living cells. Knowledge about how they interact can lead to better understanding of a great many diseases and it can help in the design of appropriate therapies.

Often studies of protein-protein interaction generate huge data sets. In the letter in Nature that was mentioned above, Jeong, Mason, Barabasi and Oltvai (2001) examined a data matrix that contained interactions linking 2114 proteins contained in yeast. Earlier experimental work had demonstrated that some of the protein molecules in yeast were lethal; if they were removed the yeast would die. Removing others, however, had no such dramatic effect. So Jeong et al. examined the question of whether the structural properties of those proteins, in particular their degree centralities, could predict which proteins were lethal and which ones were not. Their results showed that proteins of high degree were far more likely to be lethal than those of lower degree.

Subsequent articles (cited below) questioned these results. The argument was that gaps in the data called the whole analysis into question.

Usage

protein

Format

igraph object

Source

http://moreno.ss.uci.edu/data.html#pro-pro

References

Jeong, H., S. P. Mason, A.-L. Barabasi and Z. N. Oltvai. (2001). "Lethality and centrality in protein networks." Nature 411(6833): 41-42.

S. Coulomb, M. Bauer, D. Bernard, and M.-C. Marsolier-Kergoat. (2005). "Gene essentiality and the topology of protein interaction networks", Proceedings of the Royal Society B: Biological Sciences, Volume 272, Number 1573:1721-1725.

J-D. Han, D. Dupuy, N. Bertin, M. E. Cusick, and M. Vidal. (2005). "Effect of sampling on topology predictions of protein-protein interaction networks", Nature Biotechnology 23 (7):839-844.

M. Stumpf, C. Wiuf, and R. May. (2005). "Subnets of scale-free networks are not scale-free: Sampling properties of networks", PNAS 102 (12):4221-4224.